+++ test/log_test.py
@@ -101,17 +101,23 @@ class TestReadLogFile(object):
     def test_read_all(self, log_file, lines):
         log_lines = read_log_file(log_file, start_line=0, end_line=None)
 
-        assert log_lines == "".join(lines)
+        assert log_lines["logs"] == "".join(lines)
+        assert log_lines["start_line"] == 0
+        assert log_lines["end_line"] == 10
 
     def test_read_tail(self, log_file, lines):
         log_lines = read_log_file(log_file, start_line=-7, end_line=None)
 
-        assert log_lines == "".join(lines[3:])
+        assert log_lines["logs"] == "".join(lines[3:])
+        assert log_lines["start_line"] == 3
+        assert log_lines["end_line"] == 10
 
     def test_read_range(self, log_file, lines):
         log_lines = read_log_file(log_file, start_line=1, end_line=4)
 
-        assert log_lines == "".join(lines[1:4])
+        assert log_lines["logs"] == "".join(lines[1:4])
+        assert log_lines["start_line"] == 1
+        assert log_lines["end_line"] == 4
 
 
 class TestDeprecated(object):