+++ CHANGELOG.rst
@@ -1,6 +1,12 @@
 Brewtils Changelog
 ==================
 
+TBD
+------
+TBD
+
+- Updated Logs read to return dictionary with logs, start line, and end line in output object
+
 3.30.0
 ------
 9/23/25